+++ engine/src/cmd/jump_capable.cpp
@@ -271,7 +271,7 @@ bool JumpCapable::AutoPilotToErrorMessage(const Unit *target,
             failuremessage = configuration()->graphics_config.hud.already_near_message;
             return false;
         }
-        unit->ftl_energy.Deplete(true, totpercent * unit->jump.insysenergy);
+        unit->ftl_energy.Deplete(true, static_cast<double>(totpercent) * unit->jump.insysenergy);
         if (unsafe == false && totpercent == 0) {
             end = endne;
         }
